requirements.txt
celery==4.4.7
Django==3.2.7
redis==3.5.3
flower==0.9.7
django-environ==0.7.0
psycopg2-binary==2.9.1
django-celery-email==3.0.0
argh==0.26.2
watchdog==1.0.2
PyYAML==5.3.1
beautifulsoup4==4.11.1
requests==2.28.1
lxml==4.5.2